Conflicts:
	nts.uk/pr/screen/nts.uk.pr.screen.app/src/main/java/nts/uk/ctx/pr/screen/app/query/paymentdata/processor/GetPaymentDataQueryProcessor.java

+++ nts.uk/pr/proto/nts.uk.pr.proto.dom/src/main/java/nts/uk/ctx/pr/proto/dom/personalinfo/commute/PersonalCommuteValue.java
@@ -22,7 +22,7 @@ public class PersonalCommuteValue {
 	 * Commute allowance
 	 */
 	@Getter
-	private BigDecimal commuteAllowance;
+	private double commuteAllowance;
 
 	/**
 	 * Commute means attribute
@@ -46,7 +46,7 @@ public class PersonalCommuteValue {
 			int payStartYm, UseOrNot useOrNot) {
 		super();
 		this.commuteCycle = commuteCycle;
-		this.commuteAllowance = commuteAllowance;
+		this.commuteAllowance = commuteAllowance.doubleValue();
 		this.commuteMeansAttribute = commuteMeansAttribute;
 		this.payStartYm = payStartYm;
 		this.useOrNot = useOrNot;
